+++ android/app/src/main/java/com/mohammadalshouha/labourrights/MainActivity.java
@@ -1,0 +1,5 @@
+package com.mohammadalshouha.labourrights;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}